More documentation added. Better name choice for the class since it will be used as a Data-Store in a sense, not just the input reader.
